Extract shared BeanPropertyBinder code
