+++ runtime/src/main/java/io/micronaut/jackson/bind/JacksonBeanPropertyBinder.java
@@ -29,12 +29,10 @@ import io.micronaut.core.convert.exceptions.ConversionErrorException;
 import io.micronaut.core.naming.NameUtils;
 import io.micronaut.core.type.Argument;
 import io.micronaut.core.util.CollectionUtils;
-import io.micronaut.core.util.StringUtils;
 import io.micronaut.jackson.JacksonConfiguration;
+import io.micronaut.runtime.bind.AbstractBeanPropertyBinder;
 
 import javax.inject.Singleton;
-import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -48,7 +46,7 @@ import java.util.Set;
  */
 @Requires(property = "jackson.bean-property-binder.enabled", value = "true")
 @Singleton
-public class JacksonBeanPropertyBinder implements BeanPropertyBinder {
+public class JacksonBeanPropertyBinder extends AbstractBeanPropertyBinder implements BeanPropertyBinder {
 
     private final ObjectMapper objectMapper;
     private final int arraySizeThreshhold;
@@ -65,7 +63,7 @@ public class JacksonBeanPropertyBinder implements BeanPropertyBinder {
     @Override
     public BindingResult<Object> bind(ArgumentConversionContext<Object> context, Map<CharSequence, ? super Object> source) {
         try {
-            Map objectNode = buildSourceMapNode(source.entrySet());
+            Map objectNode = buildMapObject(source.entrySet());
             TypeFactory typeFactory = objectMapper.getTypeFactory();
             JavaType javaType = JacksonConfiguration.constructType(context.getArgument(), typeFactory);
             Object result = objectMapper.convertValue(objectNode, javaType);
@@ -94,7 +92,7 @@ public class JacksonBeanPropertyBinder implements BeanPropertyBinder {
     @Override
     public <T2> T2 bind(Class<T2> type, Set<? extends Map.Entry<? extends CharSequence, Object>> source) throws ConversionErrorException {
         try {
-            Map objectNode = buildSourceMapNode(source);
+            Map objectNode = buildMapObject(source);
             return objectMapper.convertValue(objectNode, type);
         } catch (Exception e) {
             throw newConversionError(null, e);
@@ -104,7 +102,7 @@ public class JacksonBeanPropertyBinder implements BeanPropertyBinder {
     @Override
     public <T2> T2 bind(T2 object, ArgumentConversionContext<T2> context, Set<? extends Map.Entry<? extends CharSequence, Object>> source) {
         try {
-            Map objectNode = buildSourceMapNode(source);
+            Map objectNode = buildMapObject(source);
             JsonNode jacksonTree = objectMapper.valueToTree(objectNode);
             objectMapper.readerForUpdating(object).readValue(jacksonTree);
         } catch (Exception e) {
@@ -116,7 +114,7 @@ public class JacksonBeanPropertyBinder implements BeanPropertyBinder {
     @Override
     public <T2> T2 bind(T2 object, Set<? extends Map.Entry<? extends CharSequence, Object>> source) throws ConversionErrorException {
         try {
-            Map objectNode = buildSourceMapNode(source);
+            Map objectNode = buildMapObject(source);
             JsonNode jacksonTree = objectMapper.valueToTree(objectNode);
             return objectMapper.readerForUpdating(object).readValue(jacksonTree);
         } catch (Exception e) {
@@ -171,152 +169,13 @@ public class JacksonBeanPropertyBinder implements BeanPropertyBinder {
         }
     }
 
-    private Map<String, Object> buildSourceMapNode(Set<? extends Map.Entry<? extends CharSequence, Object>> source) {
-        Map<String, Object> rootNode = new LinkedHashMap<>();
-        for (Map.Entry<? extends CharSequence, ? super Object> entry : source) {
-            Object value = entry.getValue();
-            String property = correctKey(entry.getKey().toString());
-            String[] tokens = property.split("\\.");
-            Object current = rootNode;
-            String index = null;
-            for (int i = 0; i < tokens.length; i++) {
-                String token = tokens[i];
-                int j = token.indexOf('[');
-                if (j > -1 && token.endsWith("]")) {
-                    index = token.substring(j + 1, token.length() - 1);
-                    token = token.substring(0, j);
-                }
-                if (i == tokens.length - 1) {
-                    if (current instanceof Map) {
-                        Map mapNode = (Map) current;
-                        if (index != null) {
-                            if (StringUtils.isDigits(index)) {
-                                int arrayIndex = Integer.parseInt(index);
-                                List arrayNode = getOrCreateListNodeAtKey(mapNode, token);
-                                arrayNode.add(arrayIndex, processValue(value));
-                            } else {
-                                Map map = getOrCreateMapNodeAtKey(mapNode, token);
-                                map.put(index, processValue(value));
-                            }
-                            index = null;
-                        } else {
-                            mapNode.put(token, processValue(value));
-                        }
-                    } else if (current instanceof List && index != null) {
-                        List arrayNode = (List) current;
-                        int arrayIndex = Integer.parseInt(index);
-                        Map mapNode = getOrCreateMapNodeAtIndex(arrayNode, arrayIndex);
-                        mapNode.put(token, processValue(value));
-                        index = null;
-                    }
-                } else {
-                    if (current instanceof Map) {
-                        Map objectNode = (Map) current;
-                        if (index != null) {
-                            if (StringUtils.isDigits(index)) {
-                                int arrayIndex = Integer.parseInt(index);
-                                List arrayNode = getOrCreateListNodeAtKey(objectNode, token);
-                                current = getOrCreateMapNodeAtIndex(arrayNode, arrayIndex);
-                            } else {
-                                Map mapNode = getOrCreateMapNodeAtKey(objectNode, token);
-                                current = getOrCreateMapNodeAtKey(mapNode, index);
-                            }
-                            index = null;
-                        } else {
-                            current = getOrCreateMapNodeAtKey(objectNode, token);
-                        }
-                    } else if (current instanceof List && StringUtils.isDigits(index)) {
-                        int arrayIndex = Integer.parseInt(index);
-                        Map jsonNode = getOrCreateMapNodeAtIndex((List) current, arrayIndex);
-
-                        current = new LinkedHashMap<>();
-                        jsonNode.put(token, current);
-                        index = null;
-                    }
-                }
-            }
-        }
-        return rootNode;
+    @Override
+    protected Object convertUnknownValue(Object o) {
+        return objectMapper.valueToTree(o);
     }
 
-    private Map getOrCreateMapNodeAtIndex(List arrayNode, int arrayIndex) {
-        if (arrayIndex >= arrayNode.size()) {
-            arrayNode = expandArrayToThreshold(arrayIndex, arrayNode);
-        }
-        Object jsonNode = arrayNode.get(arrayIndex);
-        if (!(jsonNode instanceof Map)) {
-            jsonNode = new LinkedHashMap<>();
-            arrayNode.set(arrayIndex, jsonNode);
-        }
-        return (Map) jsonNode;
+    @Override
+    protected int getArraySizeThreshold() {
+        return arraySizeThreshhold;
     }
-
-    private Map getOrCreateMapNodeAtKey(Map objectNode, Object key) {
-        Object jsonNode = objectNode.get(key);
-        if (!(jsonNode instanceof Map)) {
-            jsonNode = new LinkedHashMap<>();
-            objectNode.put(key, jsonNode);
-        }
-        return (Map) jsonNode;
-    }
-
-    private List getOrCreateListNodeAtKey(Map objectNode, Object key) {
-        Object jsonNode = objectNode.get(key);
-        if (!(jsonNode instanceof List)) {
-            jsonNode = new ArrayList<>();
-            objectNode.put(key, jsonNode);
-        }
-        return (List) jsonNode;
-    }
-
-    private List expandArrayToThreshold(int arrayIndex, List arrayNode) {
-        if (arrayIndex < arraySizeThreshhold) {
-            ArrayList arrayListNode;
-            if (arrayNode instanceof ArrayList) {
-                arrayListNode = (ArrayList) arrayNode;
-            } else {
-                arrayListNode = new ArrayList(arrayNode);
-            }
-            while (arrayNode.size() != arrayIndex + 1) {
-                arrayNode.add(arrayIndex, null);
-            }
-            return arrayListNode;
-        }
-        return arrayNode;
-    }
-
-    private Object processValue(Object o) {
-        if (o instanceof List) {
-            return processList((List) o);
-        } else if (o instanceof Map) {
-            return processMap((Map) o);
-        } else if (o instanceof Number || o instanceof String || o instanceof Boolean) {
-            return o;
-        }
-        // Not a JSON object (Map, List, Number, String, Boolean) -> needs to be converted to and processed
-        return processValue(objectMapper.valueToTree(o));
-    }
-
-    private Map processMap(Map<?, ?> map) {
-        Map newMap = new LinkedHashMap(map.size());
-        for (Map.Entry entry : map.entrySet()) {
-            Object key = correctKey(entry.getKey().toString());
-            Object value = processValue(entry.getValue());
-            newMap.put(key, value);
-        }
-        return newMap;
-    }
-
-    private List processList(List list) {
-        List newList = new ArrayList(list.size());
-        for (Object o : list) {
-            newList.add(processValue(o));
-        }
-        return newList;
-    }
-
-    private String correctKey(String key) {
-        return NameUtils.decapitalize(NameUtils.dehyphenate(key));
-    }
-
 }
